+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 ## [Unreleased]
 ### Added    
 * `cases` kwarg to `VarStat` to allow specifying which cases to use to calculate the stat
+* `cases` kwarg to `dvars_sensitivity` to allow specifying which cases to use for the sensitivity analysis
 ### Changed    
 * Switch varstatbootstrap method from 'basic' to 'BCa'
 ### Removed